moving prompts/constrained-output to batteries from core
