+++ packages/docs/docs/guides/prompting.md
@@ -44,7 +44,7 @@ The problem is that the model is predicting that a question about one state is o
 
 ## Primitives to get you prompting faster
 
-We have included a small set of prompts that we found useful in (`packages/ai-jsx/src/core/prompts.tsx`).
+We have included a small set of prompts that we found useful in [`packages/ai-jsx/src/batteries/prompts.tsx`](https://github.com/fixie-ai/ai-jsx/blob/main/packages/ai-jsx/src/batteries/prompts.tsx).
 You can use them either as shortcuts, or as a starting point if you are new to prompting:
 
 ```tsx
@@ -61,7 +61,7 @@ function App() {
 
 ## Getting the AI to say the right thing: constrained output
 
-Sometimes you just want the model to respond in a certain format (e.g. JSON or YAML), but doing so reliably can be hard.
+Sometimes you want the model to respond in a certain format (e.g. JSON or YAML), but doing so reliably can be hard.
 We provide some primitives to help with that:
 
 ```tsx
@@ -87,4 +87,4 @@ function App() {
 ```
 
 Under the hood, this model will use a combination of prompting, validating the output, and asking them the model to retry
-if the validation fails (refer to `packages/ai-jsx/src/core/constrained-output.tsx`).
+if the validation fails (refer to [`packages/ai-jsx/src/batteries/constrained-output.tsx`](https://github.com/fixie-ai/ai-jsx/blob/main/packages/ai-jsx/src/batteries/constrained-output.tsx)).
