+++ packages/examples/package.json
