+++ packages/examples/src/validated-generation.tsx
@@ -1,6 +1,6 @@
 import * as LLMx from 'ai-jsx';
 import { UserMessage } from 'ai-jsx/core/completion';
-import { JsonChatCompletion, YamlChatCompletion } from 'ai-jsx/core/constrained-output';
+import { JsonChatCompletion, YamlChatCompletion } from 'ai-jsx/batteries/constrained-output';
 import { showInspector } from 'ai-jsx/core/inspector';
 
 function App() {
